fix: 2022.2.2 hang the loading components dialog
Deadlock occur idea hang when it's opening project

BREAKING CHANGE: The wishes launcher is no longer maintained because jlib has not found a suitable configuration in kotlin 1.7. There will be no major adjustments to the wishes launcher in the future, unless it is also considered to be re-shined in the v2.0 version.

Closes #87